requirements.txt
pythondialog
data-chan
pyqtgraph
boto3
rpy2
cffi
pyQt5
statsmodels
seaborn
